Fix gpu/tests rename compile errors using CMake
It seems like the gpu/test directory was renamed to gpu/tests without
adjusting some make files and certain include paths. This caused
the gpu/tests to not build anymore. This fixes this problem.
